refactor tokengen and remove txgen as CLI
