+++ cmd/tokengen/cobra/artifactgen/gen/gen.go
