+++ cmd/tokengen/cobra/artifactgen/utils.go
