+++ cmd/tokengen/cobra/pp/common/common.go
