+++ cmd/tokengen/cobra/pp/dlog/update.go
@@ -11,7 +11,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/cmd/pp/common"
+	"github.com/hyperledger-labs/fabric-token-sdk/cmd/tokengen/cobra/pp/common"
+
 	"github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/crypto"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
