+++ cmd/tokengen/cobra/pp/fabtoken/gen.go
@@ -11,8 +11,9 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/cmd/pp/cc"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/cmd/pp/common"
+	"github.com/hyperledger-labs/fabric-token-sdk/cmd/tokengen/cobra/pp/cc"
+	"github.com/hyperledger-labs/fabric-token-sdk/cmd/tokengen/cobra/pp/common"
+
 	"github.com/hyperledger-labs/fabric-token-sdk/token/core/fabtoken"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
