+++ cmd/tokengen/cobra/pp/gen.go
@@ -7,8 +7,8 @@ SPDX-License-Identifier: Apache-2.0
 package pp
 
 import (
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/cmd/pp/dlog"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/cmd/pp/fabtoken"
+	"github.com/hyperledger-labs/fabric-token-sdk/cmd/tokengen/cobra/pp/dlog"
+	"github.com/hyperledger-labs/fabric-token-sdk/cmd/tokengen/cobra/pp/fabtoken"
 	"github.com/spf13/cobra"
 )
