+++ cmd/tokengen/cobra/version/version.go
